- remove SyftObjectTable class - move SyftMigrationRegistry to SyftObject - move migrate_to method to SyftObject - add method to get migration transform for given version
