+++ packages/syft/src/syft/types/base.py
@@ -1,3 +1,5 @@
+# stdlib
+
 # third party
 from pydantic import BaseModel
